BOL-35 refactor: 코드 위치 변경 및 자잘한 수정
