+++ adapter-out/rdb/src/main/kotlin/com/yapp/bol/group/member/MemberCommandRepositoryImpl.kt
@@ -10,5 +10,3 @@ internal class MemberCommandRepositoryImpl(
         memberRepository.save(member.toEntity())
     }
 }
-
-fun Member.toEntity(): MemberEntity = MemberEntity(id, userId, role, nickname)
